flask_demo_app_end/requirements.txt
Flask-SQLAlchemy==3.1.1
psycopg2==2.9.9